requirements.txt
beautifulsoup4==4.10.0
black==22.1.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
idna==3.3
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.0
pushbullet.py==0.12.0
pymongo==4.0.1
python-dotenv==0.19.2
python-magic==0.4.25
requests==2.27.1
soupsieve==2.3.1
tomli==2.0.1
typing-extensions==4.1.0
urllib3==1.26.8
websocket-client==1.2.3
